requirements.txt
# DW-Bench dependencies
torch>=2.0
torch_geometric>=2.5
networkx>=3.0
openai>=1.0
requests>=2.28
sentence-transformers>=2.2
faiss-cpu>=1.7
numpy>=1.24
pandas>=1.5
python-dotenv>=1.0
